+++ composeApp/build.gradle.kts
@@ -42,7 +42,7 @@ kotlin {
 
             implementation(libs.ktor.client.okhttp)
             implementation(libs.play.services.code.scanner)
-            implementation(libs.core)
+            implementation(libs.zxing)
         }
         iosMain.dependencies {
             implementation(libs.ktor.client.darwin)
