+++ composeApp/src/androidMain/kotlin/com/github/familyvault/DI.android.kt
@@ -4,16 +4,13 @@ import com.github.familyvault.backend.client.IPrivMxClient
 import com.github.familyvault.backend.client.createPrivMxClient
 import com.github.familyvault.database.AppDatabase
 import com.github.familyvault.database.createAppDatabase
-import com.github.familyvault.services.IQRCodeGenerationService
-import com.github.familyvault.services.IQRCodeScannerService
-import com.github.familyvault.services.createQrCodeGenerationService
-import com.github.familyvault.services.createQrCodeScannerService
+import com.github.familyvault.services.IQRCodeService
+import com.github.familyvault.services.createQrCodeService
 import org.koin.dsl.bind
 import org.koin.dsl.module
 
 actual fun getPlatformModules() = module {
     single { createPrivMxClient(get()) }.bind<IPrivMxClient>()
     single { createAppDatabase(get()) }.bind<AppDatabase>()
-    single { createQrCodeScannerService(get()) }.bind<IQRCodeScannerService>()
-    single { createQrCodeGenerationService(get()) }.bind<IQRCodeGenerationService>()
+    single { createQrCodeService(get()) }.bind<IQRCodeService>()
 }
