--- composeApp/src/androidMain/kotlin/com/github/familyvault/services/createQrCodeGenerationService.kt
@@ -1,7 +1,0 @@
-package com.github.familyvault.services
-
-import android.content.Context
-
-fun createQrCodeGenerationService(context: Context): IQRCodeGenerationService {
-    return QrCodeGeneratorService(context)
-}
