--- composeApp/src/androidMain/kotlin/com/github/familyvault/services/createQrCodeScannerService.kt
@@ -1,7 +1,0 @@
-package com.github.familyvault.services
-
-import android.content.Context
-
-fun createQrCodeScannerService(context: Context): IQRCodeScannerService {
-    return QRCodeScannerService(context)
-}
