+++ composeApp/src/androidMain/kotlin/com/github/familyvault/services/createQrCodeService.kt
@@ -1,0 +1,7 @@
+package com.github.familyvault.services
+
+import android.content.Context
+
+fun createQrCodeService(context: Context): IQRCodeService {
+    return QRCodeService(context)
+}
