+++ composeApp/src/commonMain/kotlin/com/github/familyvault/screens/start/DebugScreenContextId.kt
@@ -8,21 +8,20 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.semantics.Role.Companion.Image
 import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.currentOrThrow
 import com.github.familyvault.components.InitialScreenButton
 import com.github.familyvault.screens.main.MainScreen
 import com.github.familyvault.services.IFamilyGroupSessionService
-import com.github.familyvault.services.IQRCodeGenerationService
+import com.github.familyvault.services.IQRCodeService
 import org.koin.compose.koinInject
 
 class DebugScreenContextId : Screen {
     @Composable
     override fun Content() {
         val familyGroupSessionService = koinInject<IFamilyGroupSessionService>()
-        val qrCodeGenerationService = koinInject<IQRCodeGenerationService>()
+        val qrCodeGenerationService = koinInject<IQRCodeService>()
         val navigator = LocalNavigator.currentOrThrow
         val contextId = familyGroupSessionService.getContextId()
         Column(
