--- composeApp/src/commonMain/kotlin/com/github/familyvault/services/IQRCodeGenerationService.kt
@@ -1,7 +1,0 @@
-package com.github.familyvault.services
-
-import androidx.compose.ui.graphics.ImageBitmap
-
-interface IQRCodeGenerationService {
-    fun generateQRCode(qrCodeContent: String): ImageBitmap?
-}
