+++ composeApp/src/commonMain/kotlin/com/github/familyvault/services/IQRCodeService.kt
@@ -1,7 +1,9 @@
 package com.github.familyvault.services
 
+import androidx.compose.ui.graphics.ImageBitmap
 import com.github.familyvault.models.QrCodeScanResponse
 
-interface IQRCodeScannerService {
+interface IQRCodeService {
     suspend fun scanQRCode(): QrCodeScanResponse
+    fun generateQRCode(qrCodeContent: String): ImageBitmap?
 }
