+++ gradle/libs.versions.toml
@@ -5,7 +5,7 @@ android-minSdk = "24"
 android-targetSdk = "34"
 androidx-activityCompose = "1.9.3"
 compose-multiplatform = "1.7.0"
-core = "3.4.0"
+zxing = "3.4.0"
 kotlin = "2.0.21"
 playServicesCodeScanner = "16.1.0"
 privmxEndpoint = "2.1.0"
@@ -22,7 +22,7 @@ sqlite = "2.5.0-SNAPSHOT"
 
 [libraries]
 androidx-activity-compose = { module = "androidx.activity:activity-compose", version.ref = "androidx-activityCompose" }
-core = { module = "com.google.zxing:core", version.ref = "core" }
+zxing = { module = "com.google.zxing:core", version.ref = "zxing" }
 play-services-code-scanner = { module = "com.google.android.gms:play-services-code-scanner", version.ref = "playServicesCodeScanner" }
 voyager-transitions = { module = "cafe.adriel.voyager:voyager-transitions", version.ref = "voyagerVersion" }
 voyager-tab-navigator = { module = "cafe.adriel.voyager:voyager-tab-navigator", version.ref = "voyagerVersion" }
